seoul: add intel82576vf for 64bit

+++ ports/src/vancouver/nova_user_env.cc
@@ -115,7 +115,7 @@ void *operator new[](size_t size)
 
 void *operator new[](size_t size, Aligned const alignment)
 {
-	unsigned int align = alignment.alignment;
+	size_t align = alignment.alignment;
 	void *res = heap_alloc(size + align);
 	if (res)
 		Genode::memset(res, 0, size + align);
